requirements.txt
langchain==0.3.25
langchain-core==0.3.59
langchain-community==0.3.24
langchain-openai==0.3.17
langchain-anthropic==0.3.12
python-dotenv==1.0.1
pydantic==2.11.3
wikipedia==1.4.0
duckduckgo-search